add storage_map
